backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pypdf==4.3.1
httpx==0.27.2    # <<< pin to avoid 'proxies' error
cohere==5.9.0
pytesseract==0.3.13
pillow==10.4.0